New Player Shop Overhaul
lowered prices and added to stock of the bouldertanks' shop. new items were also added; basic +1 - +4 AC clothes and boots that do not have a UMD requirement, as well as 4 new trinkets (not wands) that cast death ward, prot from align, restoration, and shield for 5 charges a cast so as to enable new players' progression without stepping on crafters' toes.