frenzy buffs
removed AC and DOT penalties, buffed the offensive capabilities of the frenzy through hasting for the duration since FB is trading defense for offense, and gave crit immunity for four rounds upon activation of the ability so as to keep the FB from going 100 to 0 in crowd scenarios before they ever get started. FB's core class feature should be beneficial instead of overwhelmingly detrimental with such in place!