Mageslayer tweak
Removed feat script and turned over management of mageslayer damage bonus to a function in class_mageslayer_utils that will remove and then re-calculate and apply the damage bonus on login and level up to prevent double bonus if player re-levels. Added that call to login and level up scripts.