Fix for mageslayer on-hit, + xtras
Fixed mage slayer on hit so that vanilla mage slayer can use any on-hit property of any weapon or shield but not activated abilities such as itempropertycastspell(), the other 2 can use both but only arcane or divine, depending on path. Also gave vanilla mage slayers access to "trinkets" that don't require magical skill and the other 2 access to the ones that aligne with path. Streamlined code, added controls via constants for trinkets