Update x2_inc_craft.nss
Sub-spells (like Prot from Evil is a sub-spell of prot from align) don't always have their spell level for class listed in the 2da because it's unnecessary so long as the master spell gets theirs. This was causing some spells to be unrecognized when scribing/brewing/wanding. Added logic to search for a master spell and get its spell level for class instead of subspell.